+++ src/vibey_tools/gh/vibey_gh/interfaces/__init__.py
@@ -1,6 +1,7 @@
 # Made with ❤️ by [Vibey](https://the-vibey-project.github.io/vibey/), Developed by [Adam Matthew Steinberger](https://vibewithadam.matthewsteinberger.com/) ([@adammatthewsteinberger](https://github.com/adammatthewsteinberger/)).
 """Seams this package declares. Interfaces declare; they never consume."""
 
+from vibey_gh.interfaces.book_interface import MainExtractorInterface
 from vibey_gh.interfaces.chapter_sanitizer_interface import ChapterSanitizerInterface
 
-__all__ = ["ChapterSanitizerInterface"]
+__all__ = ["ChapterSanitizerInterface", "MainExtractorInterface"]